+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Unreleased
 
 - add --install option to test command
 - allow backslash line continuations in recipe files
+- fix missing ``/ocpp/csms`` route for charger details
+- add ``--all`` flag to ``web.app.setup_app`` to auto-load sub-projects and
+  resolve sub-project views without wrappers
 
 0.4.51 [build edbd5b]
 ---------------------